Deduplicate post display code and add consistent swipe gestures
- Create unified PostDisplayView replacing PostHeaderView and PostRowView content
- Add swipe gestures and context menu to PostHeaderView matching FeedView behavior
- Move shared components (PostContextMenu, ThumbnailView) to PostDisplayView.swift
- Fix FeedView tap area issues by conditionally applying tap gestures
- Ensure consistent UI/UX between feed and comments views
- Reduce code duplication and improve maintainability
